# feat: Support lambda expression compile and delegate cache 